requirements.txt
python-dotenv
langchain
langchain_community
langchain-pinecone
google-cloud-aiplatform
langchain-google-vertexai
beautifulsoup4
langchain-openai
